signed/unsigned mismatch
git-svn-id: svn://www.xmlblaster.org/xmlBlaster/trunk@14446 00fa6119-6a0f-0410-b6da-8b96a2b62773

+++ xmlBlaster/testsuite/src/c/TestStress.c
@@ -152,17 +152,17 @@ static const char * test_stress()
 
    for (iWait=0; iWait<10; iWait++) {
       printf("[client] Publish of %d messages success, received %d updates\n", numPublish, (int32_t)updateCounter);
-      if (updateCounter >= numPublish)
+      if ((int)updateCounter >= numPublish)
          break;
       sleepMillis(500);
    }
 
    mu_assert("No update arrived", *updateContent != '\0');
-   if (updateCounter < numPublish) {
+   if ((int)updateCounter < numPublish) {
       freeXmlBlasterAccessUnparsed(xa);
       mu_assert("Missing updates", updateCounter == numPublish);
    }
-   else if (updateCounter > numPublish) {
+   else if ((int)updateCounter > numPublish) {
       printf("[client] WARN: Publish of %d messages but received %d updates\n", numPublish, (int32_t)updateCounter);
    }
    printf("[client] updateContent = %s, CONTENT = %s\n", updateContent, CONTENT);
